requirements.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.2
django-crispy-forms==1.6.1
django-hosts==2.0
django-pagedown==0.1.3
django-storages==1.6.5
django-toolbelt==0.0.1
gunicorn==19.7.1
Pillow==4.2.1
psycopg2==2.7.3
whitenoise==3.3.0
